+++ OS_Lab1/Gx.h
@@ -4,7 +4,7 @@ class Gx : public Function
 {
 public:
 	Gx();
-	Result Compute(int x, const std::atomic<bool>& flag) override;
+	Result Compute(float x, const std::atomic<bool>& flag) override;
 	void SetFunctionName() override;
 };
